another attempt to get the latest multiprocessing changes to work on windows
